fix: use string for extraInitContainers
The object/string type mismatch prevents this being used in some cases
(#424). This pattern affects some other keys but I didn't want to make
too big a change as I don't know how to test all of this carefully, and
the CI doesn't seem to have coverage of individual values.

I tried regenerating the docs with helm-docs but that changed lots of
unrelated values, so I limited the changes to the extraInitContainers
value.